split up in multiple files and add Instrument reference instead of name
