+++ qcodes/instrument/ip_to_visa.py
@@ -1,12 +1,15 @@
-from qcodes.instrument.base import Instrument, InstrumentLoggerAdapter
+import logging
+from qcodes.instrument.base import Instrument
 from qcodes.instrument.ip import IPInstrument
 # previous to introducing the `InstrumentLoggerAdapter` the IPToVisa instrument
 # was logging in the name of the `VisaInstrument`. To maintain that behaviour
 # import the `instrument.visa.log` and log to this one.
-from qcodes.instrument.visa import VisaInstrument, log
+from qcodes.instrument.visa import VisaInstrument, log, VISA_LOGGER
 from qcodes.instrument_drivers.american_magnetics.AMI430 import AMI430
 from qcodes.utils.helpers import strip_attrs
+from qcodes.logger.instrument_logger import get_instrument_logger
 import qcodes.utils.validators as vals
+
 
 # This module provides a class to make an IPInstrument behave like a
 # VisaInstrument. This is only meant for use with the PyVISA-sim backend
@@ -47,9 +50,8 @@ class IPToVisa(VisaInstrument, IPInstrument): # type: ignore
 
 
         Instrument.__init__(self, name, metadata=metadata, **newkwargs)
+        self.visa_log = get_instrument_logger(self, VISA_LOGGER)
 
-        self.log = InstrumentLoggerAdapter(logger=log,
-                                           extra={'instrument': self.full_name})
         ##################################################
         # __init__ of VisaInstrument
