+++ qcodes/logger/logger.py
@@ -11,16 +11,11 @@ from copy import copy
 
 from typing import Optional, List, Union, Sequence
 
-import pandas
-from pandas.core.series import Series
-
 import qcodes as qc
-from qcodes.instrument.base import InstrumentFilter, InstrumentBase
-
-LevelType = Union[int, str]
-
 
 log = logging.getLogger(__name__)
+
+LevelType = Union[int, str]
 
 LOGGING_DIR = "logs"
 LOGGING_SEPARATOR = ' ¦ '
@@ -32,6 +27,7 @@ FORMAT_STRING_DICT = OrderedDict([
     ('asctime', 's'),
     ('name', 's'),
     ('levelname', 's'),
+    ('module', 's'),
     ('funcName', 's'),
     ('lineno', 'd'),
     ('message', 's')])
@@ -199,109 +195,7 @@ def start_all_logging() -> None:
     start_command_history_logger()
 
 
-def log_to_dataframe(log: List[str]=None,
-                     columns: Optional[List[str]]=None,
-                     separator: Optional[str]=None) -> pandas.DataFrame:
-    """
-    Return the provided or default log string as a pandas DataFrame.
 
-    Unless `qcodes.utils.logger.LOGGING_SEPARATOR` or
-    `qcodes.utils.logger.FORMAT_STRING...` have been changed using the default
-    for the columns and separtor arguments is encouraged.
-
-    Lines starting with a digit are ignored. In the log setup of `start_logging`
-    Traceback messages are also logged. These start with a digit.
-
-    Args:
-        log: log content
-        columns: column headers for the returned dataframe, defaults to
-            columns used by handlers set up by `start_logger`.
-        separator: seperator of the logfile to seperate the columns, defaults to
-            separtor used by handlers set up by `start_logger`.
-
-    Retruns:
-        Pandas DataFrame containing the log content.
-    """
-    separator = separator or LOGGING_SEPARATOR
-    columns = columns or list(FORMAT_STRING_DICT.keys())
-    # note: if we used commas as separators, pandas read_csv
-    # could be faster than this string comprehension
-
-    split_cont = [line.split(separator) for line in log
-                  if line[0].isdigit()]  # avoid tracebacks
-    dataframe = pandas.DataFrame(split_cont, columns=columns)
-
-    return dataframe
-
-
-def logfile_to_dataframe(logfile: Optional[str]=None,
-                         columns: Optional[List[str]]=None,
-                         separator: Optional[str]=None) -> pandas.DataFrame:
-    """
-    Return the provided or default logfile as a pandas DataFrame.
-
-    Unless `qcodes.utils.logger.LOGGING_SEPARATOR` or
-    `qcodes.utils.logger.FORMAT_STRING...` have been changed using the default
-    for the columns and separtor arguments is encouraged.
-
-    Lines starting with a digit are ignored. In the log setup of `start_logging`
-    Traceback messages are also logged. These start with a digit.
-
-    Args:
-        logfile: name of the logfile; defaults to current default log file.
-        columns: column headers for the returned dataframe, defaults to
-            columns used by handlers set up by `start_logger`.
-        separator: seperator of the logfile to seperate the columns, defaults to
-            separtor used by handlers set up by `start_logger`.
-
-    Retruns:
-        Pandas DataFrame containing the logfile content.
-    """
-    logfile = logfile or get_log_file_name()
-    with open(logfile) as f:
-        raw_cont = f.readlines()
-
-    return log_to_dataframe(raw_cont, columns, separator)
-
-
-def time_difference(firsttimes: Series,
-                    secondtimes: Series,
-                    use_first_series_labels: bool=True) -> Series:
-    """
-    Calculate the time differences between two series
-    containing time stamp strings as their values.
-
-    Args:
-        firsttimes: The oldest times
-        secondtimes: The youngest times
-        use_first_series_labels: If true, the returned Series
-            has the same labels as firsttimes. Else it has
-            the labels of secondtimes
-
-    Returns:
-        A Series with float values of the time difference (s)
-    """
-
-    if ',' in firsttimes.iloc[0]:
-        nfirsttimes = firsttimes.str.replace(',', '.')
-    else:
-        nfirsttimes = firsttimes
-
-    if ',' in secondtimes.iloc[0]:
-        nsecondtimes = secondtimes.str.replace(',', '.')
-    else:
-        nsecondtimes = secondtimes
-
-    t0s = nfirsttimes.astype("datetime64[ns]")
-    t1s = nsecondtimes.astype("datetime64[ns]")
-    timedeltas = (t1s.values - t0s.values).astype('float')*1e-9
-
-    if use_first_series_labels:
-        output = pandas.Series(timedeltas, index=nfirsttimes.index)
-    else:
-        output = pandas.Series(timedeltas, index=nsecondtimes.index)
-
-    return output
 
 
 @contextmanager
@@ -345,83 +239,6 @@ def console_level(level: LevelType):
     global console_handler
     with handler_level(level, handler=console_handler):
         yield
-
-
-@contextmanager
-def filter_instrument(instrument: Union[InstrumentBase,
-                                        Sequence[InstrumentBase]],
-                      handler: Optional[
-                          Union[logging.Handler,
-                                Sequence[logging.Handler]]]=None,
-                      level: Optional[LevelType]=None):
-    """
-    Context manager that adds a filter that only enables the log messages of
-    the supplied instruments to pass.
-    Example:
-        >>> h1, h2 = logger.get_console_handler(), logger.get_file_handler()
-        >>> with logger.filter_instruments((qdac, dmm2), handler=[h1, h1]):
-        >>>     qdac.ch01(1)  # logged
-        >>>     v = dmm2.v()  # not logged
-
-    Args:
-        level: level to set the handlers to
-        handler: single or sequence of handlers which to change
-    """
-    if handler is None:
-        global console_handler
-        handler = (console_handler,)
-    if isinstance(handler, logging.Handler):
-        handler = (handler,)
-
-    instrument_filter = InstrumentFilter(instrument)
-    for h in handler:
-        h.addFilter(instrument_filter)
-    try:
-        if level is not None:
-            with handler_level(level, handler):
-                yield
-        else:
-            yield
-    finally:
-        for h in handler:
-            h.removeFilter(instrument_filter)
-
-
-@contextmanager
-def capture_dataframe(level: LevelType=logging.DEBUG,
-                      logger: logging.Logger=None):
-    """
-    Context manager to capture the logs in a `pandas.DataFrame`
-
-    Example:
-        >>> with logger.capture_dataframe() as (handler, cb):
-        >>>     qdac.ch01(1)  # some commands
-        >>>     data_frame = cb()
-
-    Args:
-        level: level at which to capture
-        handler: single or sequence of handlers which to change
-
-    Returns:
-        Tuple of handler that is used to capture the log messages and callback
-        that returns the cummulative pandas.DataSet at any given
-        point (within the context)
-    """
-    # get root logger if none is specified.
-    logger = logger or logging.getLogger()
-    with io.StringIO() as log_capture:
-        string_handler = logging.StreamHandler(log_capture)
-        string_handler.setLevel(level)
-        format_string = LOGGING_SEPARATOR.join(FORMAT_STRING_ITEMS)
-        formatter = logging.Formatter(format_string)
-        string_handler.setFormatter(formatter)
-
-        logger.addHandler(string_handler)
-        try:
-            yield string_handler, lambda: log_to_dataframe(
-                log_capture.getvalue().splitlines())
-        finally:
-            logger.removeHandler(string_handler)
 
 
 class LogCapture():
